+++ harvest/lib/transform/elastic-search/index.js
@@ -26,104 +26,6 @@ const contextFile = JSON.parse(fs.readFileSync(contextPath, 'utf8'));
 const framePath = path.join(__dirname, 'frames', 'default.json');
 const frameFile = JSON.parse(fs.readFileSync(framePath, 'utf8'));
 
-/**
- * @method normalizeBooleans
- * @description Recursively mutates jsonld boolean values
- * (for example `{"@value": "true", "@type": "xsd:boolean"}`)
- * to native javascript booleans.
- * @param {@} node
- * @returns
- */
-const normalizeBooleans = (node) => {
-  if (!node || typeof node !== 'object') return;
-  Object.keys(node).forEach(k => {
-    const v = node[k];
-    if (Array.isArray(v)) {
-      v.forEach(normalizeBooleans);
-    } else if (v && typeof v === 'object') {
-      if (
-        v['@value'] !== undefined &&
-        (v['@type'] === 'xsd:boolean' || v['@type'] === 'http://www.w3.org/2001/XMLSchema#boolean')
-      ) {
-        node[k] = (v['@value'] === true || v['@value'] === 'true');
-      } else {
-        normalizeBooleans(v);
-      }
-    }
-  });
-};
-
-/**
- * @method normalizeScalars
- * @description Recursively mutates jsonld scalar values (boolean, integer, string,
- * for example `{"@value": "true", "@type": "xsd:boolean"}`)
- * to native javascript types.
- * @param {@} node
- * @returns
- */
-const normalizeScalars = (node) => {
-  if (!node || typeof node !== 'object') return;
-  for (const k of Object.keys(node)) {
-    const v = node[k];
-    if (Array.isArray(v)) v.forEach(normalizeScalars);
-    else if (v && typeof v === 'object' && '@value' in v) {
-      if (v['@type'] === 'xsd:boolean') node[k] = v['@value'] === true || v['@value'] === 'true';
-      else if (v['@type'] === 'xsd:integer') node[k] = parseInt(v['@value'], 10);
-      else node[k] = v['@value'];
-    } else if (v && typeof v === 'object') normalizeScalars(v);
-  }
-}
-
-/**
- * @method normalizeUnicodeSpacesDeep
- * @description Recursively mutates jsonld string values to replace
- * various unicode space characters with regular spaces.
- * Some descriptions in works/grants contain different types of spaces
- * which causes issues during diff comparisons.
- *
- * The following unicode space characters are targeted:
- * ```
- * \u00A0 - Non-breaking space (NBSP)
- * \u202F - Narrow no-break space
- * \u2007 - Figure space
- * \u2000-\u200A - Various em/en spaces, thin spaces, hair spaces, etc.
- * \u205F - Medium mathematical space
- * ```
- * @param {Object} graph
- * @returns {Object} updated graph
- */
-const normalizeUnicodeSpacesDeep = (graph) => {
-  const seen = new WeakSet();
-  const SPACE_RE = /[\u00A0\u202F\u2007\u2000-\u200A\u205F]/g;
-
-  const fixString = (s) => {
-    if (typeof s !== 'string') return s;
-    return s.replace(SPACE_RE, ' ');
-  };
-
-  const recurse = (node) => {
-    if (!node || typeof node !== 'object') return;
-    if (seen.has(node)) return;
-    seen.add(node);
-
-    for (const key of Object.keys(node)) {
-      const v = node[key];
-      if (typeof v === 'string') {
-        node[key] = fixString(v);
-      } else if (Array.isArray(v)) {
-        for (let i = 0; i < v.length; i++) {
-          if (typeof v[i] === 'string') v[i] = fixString(v[i]);
-          else if (v[i] && typeof v[i] === 'object') recurse(v[i]);
-        }
-      } else if (v && typeof v === 'object') {
-        recurse(v);
-      }
-    }
-  };
-
-  graph.forEach(n => recurse(n));
-  return graph;
-};
 
 /**
  * @method collapseSingleItemPrimitiveArrays
