+++ harvest/lib/transform/elastic-search/to-work-webapp.js
@@ -14,144 +14,6 @@ const contextFile = JSON.parse(fs.readFileSync(contextPath, 'utf8'));
 const framePath = path.join(__dirname, 'frames', 'default.json');
 const frameFile = JSON.parse(fs.readFileSync(framePath, 'utf8'));
 
-/**
- * @method transformWork
- * @description Given a subject, transform the corresponding work
- * 
- * @param {String} subject the subject URI of the work to transform
- * @param {Object} opts 
- */
-async function transformWork(subject, opts={}) {
-  const partitionKeys = ['year-week-'+getYearWeek(opts.date), 'ae-std'];
-  let relNodes = [];
-  let expertNodes = [];
-
-  const rdfResp = await cache.findRelatedExperts(subject, {partitionKeys});
-  let workNode = null;
-
-  // loop through all found rel files, extract the authorship nodes that relate to an expert
-  for (const res of rdfResp.results) {
-    const fp = res.filepath;
-    if (!fp) continue;
-    try {
-      const rel = JSON.parse(await cache.read(fp));
-      if ( !workNode ) {
-        workNode = getNodeByType(rel, SHORT_TYPES.WORKS, {match: true});
-      }
-
-      const items = getGraphAsItems(rel);
-
-      for (const node of items) {
-        let parsedNode = _parseWorkNode(subject, node);
-        if( !parsedNode ) continue;
-        
-        relNodes.push(parsedNode);
-        let expert = await _findWorkExpert(parsedNode, partitionKeys, subject);        
-        if( expert ) expertNodes.push(expert);
-      }
-    } catch (e) {
-      logger.error(`Failed to read/parse RDF-found rel file ${fp}`, e);
-    }
-  }
-
-
-
-  let workGraph = await jsonld.compact([workNode, ...relNodes], contextFile["@context"]);
-  workGraph = [...workGraph["@graph"], ...expertNodes];
-
-  workGraph = workGraph.map(node => {
-    if( node.rank ) node.rank = parseInt(node.rank);
-    if( node['is-visible'] ) node['is-visible'] = node['is-visible'] === 'true';
-    return node;
-  });
-
-  return workGraph;
-}
-
-async function _findWorkExpert(workRelNode, partitionKeys, subject) {
-  let expertId = asArray(workRelNode['http://vivoweb.org/ontology/core#relates'])
-    .find(value => value['@id'].startsWith('http://experts.ucdavis.edu/expert/'));
-
-  if( !expertId ) {
-    logger.warn(`No expert relates found in node ${workRelNode['@id']} for subject ${subject}`);
-    return null;
-  } 
-
-  expertId = expertId['@id'];
-  let relatedExpert = await cache.findRelatedExperts(expertId, {partitionKeys});
-  if( !relatedExpert.results.length ) {
-    logger.warn(`No related experts found for expertId ${expertId} in node ${parsedNode['@id']} for subject ${subject}`);
-    return null;
-  }
-
-  let person = JSON.parse(await cache.read(relatedExpert.results[0].filepath));
-  // console.log(SHORT_TYPES.EXPERT)
-  
-
-  const frameDoc = {
-    ...frameFile
-  };
-  frameDoc["@context"] = contextFile["@context"];
-
-  const framedRaw = await jsonld.frame(
-    person,
-    frameDoc,
-    {
-      embed: '@always',
-      omitGraph: false,
-      explicit: false,
-      requireAll: false
-    }
-  );
-    console.log(JSON.stringify(framedRaw, null, 2));
-
-
-  let compacted = await jsonld.compact(framedRaw, contextFile["@context"], {
-    embed: '@always'
-  }); 
-
-
-  person = getNodeByType(compacted, SHORT_TYPES.EXPERT, {match: true});
-
-  console.log(person);
-
-
-  return createSimplifiedExpert(person);
-}
-
-
-function _parseWorkNode(subject, node) {
-  if (!node || !node['@id']) return;
-
-  // Only consider relationship nodes
-  if (!(typeof node['@id'] === 'string' && node['@id'].includes('/relationship/'))) {
-    logger.debug(`Skipping non-relationship node ${node['@id']}`);
-    return;
-  }
-
-  const relatesAny = node['http://vivoweb.org/ontology/core#relates'] || node['ucdlib:relates-to'] || node['relatesTo'];
-  const relatesArr = Array.isArray(relatesAny) ? relatesAny : (relatesAny ? [relatesAny] : []);
-  if (!relatesArr.length) {
-    logger.warn(`Relationship node ${node['@id']} has no relates field`);
-    return;
-  }
-
-  // must reference the publication subject
-  const referencesPub = relatesArr.some(r => {
-    const rid = (typeof r === 'string') ? r : (r && r['@id'] ? r['@id'] : null);
-    return rid && rid.split('#')[0] === subject;
-  });
-  if (!referencesPub) {
-    logger.warn(`Relationship node ${node['@id']} does not reference the publication subject ${subject}`);
-    return;
-  }
-
-  // clone node and set vivoweb relates to the collected relates
-  const outNode = JSON.parse(JSON.stringify(node));
-  outNode['http://vivoweb.org/ontology/core#relates'] = relatesArr;
-  
-  return outNode;
-}
 
 /**
  * @method generateWorkFiles
@@ -359,52 +221,7 @@ function getWorkNodes(framedDocument) {
   });
 }
 
-/**
- * @method updateWorkRelatedByRelates
- * @description Update relatedBy relates fields in work nodes to use string expert references
- * Similar to grant pattern but adapted for work authorships
- * @param {*} workDocument the work document containing work and expert nodes
- */
-function updateWorkRelatedByRelates(workDocument) {
-  // Find the expert node
-  const expertNode = workDocument["@graph"].find(
-    n => n && (n["@type"] === "Expert" || (Array.isArray(n["@type"]) && n["@type"].includes("Expert")))
-  );
-  if (!expertNode) return;
 
-  const expertIdStr = expertNode['@id'];
-
-  // Find the work node
-  workDocument["@graph"].forEach(node => {
-    const types = Array.isArray(node['@type']) ? node['@type'] : [node['@type']];
-    const isWork = types.some(t =>
-      t.includes('Work') ||
-      t.includes('ScholarlyArticle') ||
-      t.includes('Article') ||
-      t.includes('Publication')
-    );
-    if (!isWork) return;
-
-    // Process relatedBy (authorships)
-    if (Array.isArray(node.relatedBy)) {
-      node.relatedBy.forEach(authorship => {
-        if (!authorship || authorship.relates === undefined) return;
-        let relatesArr = Array.isArray(authorship.relates) ? authorship.relates : [authorship.relates];
-        // Ensure string @id only; collect ids, add work + expert if present as objects without duplication
-        const ids = new Set();
-        relatesArr.forEach(r => {
-          if (typeof r === 'string') ids.add(r);
-          else if (r && typeof r === 'object' && r['@id']) ids.add(r['@id']);
-        });
-        // Guarantee work id present when authorship references work
-        if (!ids.has(node['@id'])) ids.add(node['@id']);
-        // Guarantee expert id present when authorship references expert
-        if (!ids.has(expertIdStr)) ids.add(expertIdStr);
-        authorship.relates = Array.from(ids);
-      });
-    }
-  });
-}
 
 /**
  * @method buildWorkRelatedBy
